Move navigation-rail into components/ folder.
This also moves renderLink from drawer/ to navigation-rail/.
