+++ client/src/components/navigation-rail/components/TutorialRailItem.helpers.tsx
